platform/x86: Move ideapad ACPI helpers to a new header
These functions will be used by a driver written by Gergo Koteles to
detect the tablet mode switch in Lenovo Yoga laptops. These changes were
discussed in review of that patch.

This is the minimal set of functions needed in that driver, there are
several more small functions left in the ACPI Helpers section in
ideapad-laptop.c. The only change is the functions are now marked inline
as requested in the review comments.

Link: https://lore.kernel.org/r/20221004214332.35934-1-soyer@irl.hu/
Signed-off-by: Andrew Kallmeyer <kallmeyeras@gmail.com>